refactor: Extract SAMUpdateWorker to separate module
- Move SAMUpdateWorker class from main_window.py to workers/sam_update_worker.py
- Update workers/__init__.py to export SAMUpdateWorker
- Maintain all existing functionality and signals
- Reduces main_window.py by 123 lines
- All tests pass and code quality checks clean
