+++ src/lazylabel/ui/workers/__init__.py
@@ -1,5 +1,6 @@
 """Worker thread classes for background operations."""
 
 from .image_discovery_worker import ImageDiscoveryWorker
+from .sam_update_worker import SAMUpdateWorker
 
-__all__ = ["ImageDiscoveryWorker"]
+__all__ = ["ImageDiscoveryWorker", "SAMUpdateWorker"]
